requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pyttsx3==2.90
SpeechRecognition==3.10.0
PyAudio==0.2.11
openai==1.3.5
flask-socketio==5.3.6
python-socketio==5.9.0
